Narrow fclShopMisc f25e0 parameters

+++ src/Event/mt_evtLoadSave.c
@@ -41,7 +41,7 @@ extern void FUN_004c0420_evt(void *data,int size,int count,int arg);
 #pragma alias FUN_00361ca0_evt FUN_00361ca0
 extern u32 FUN_00361ca0_evt(int param_1,u64 param_2);
 void FUN_0038fa10(int param_1,u64 param_2);
-void FUN_0038ffb0(int param_1,u64 param_2);
+void FUN_0038ffb0(int param_1,int param_2);
 void FUN_003902c0(int param_1,u64 param_2);
 void FUN_003905f0(int param_1,u64 param_2);
 void FUN_00390920(int param_1,u64 param_2);
@@ -2478,7 +2478,7 @@ void FUN_0038ffb0(int param_1,int param_2)
 // FUN_003902C0 NONMATCHING
 
 
-void FUN_003902c0(int param_1,u64 param_2)
+void FUN_003902c0(int param_1,int param_2)
